split up classes in planner_node.hpp into 3 files
